pyproject.toml
[tool.poetry]
name = "human-resource-management"
version = "1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "human_resource_management"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
django-environ = "^0.9.0"
mysqlclient = "^2.1.1"
django-ajax-datatable = "^4.4.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
